Orders ok, set has finish

+++ fixme-core/src/main/java/org/fixme/core/protocol/types/WalletObject.java
@@ -48,7 +48,7 @@ public class WalletObject extends BaseCollection implements INetworkType {
 		public void deserialize(Json buffer) {
 			this.wallet = buffer.getString("WALLET");
 			this.instrument = buffer.getString("INSTRUMENT");
-			this.quantity = buffer.getInt("QUANTITY");
+			this.quantity = buffer.getFloat("QUANTITY");
 		}
 		
 		@Override
